Add details screen, add banner autoscroll

+++ domain/src/main/java/com/books/domain/entities/Banner.kt
@@ -1,3 +1,3 @@
 package com.books.domain.entities
 
-data class Banner(val id: Long, val book_id: Long, val cover: String)
+data class Banner(val id: Int, val book_id: Int, val cover: String)
